refactor: [P4ADEV-3252] removes useClassificationSearch
